lesson_5: update types
