+++ lesson_5/src/helpers.ts
@@ -1,5 +1,7 @@
 import { ValidationError } from "express-validator";
 import { FieldError } from "./types";
+import { ParsedQs } from "qs";
+import { SortDirection } from "./constants";
 
 export const getFormattedErrors = (errors: ValidationError[]) => {
   const formattedErrors = errors.reduce<{ errorsMessages: FieldError[] }>(
@@ -18,4 +20,24 @@ export const getFormattedErrors = (errors: ValidationError[]) => {
   );
 
   return formattedErrors;
+};
+
+export const getFiltersFromQuery = (query: ParsedQs) => {
+  const pagination = {
+    pageNumber: +(query.pageNumber || 1),
+    pageSize: +(query.pageSize || 10),
+  };
+
+  const sortDirection = (): SortDirection => {
+    if (query.sortDirection === SortDirection.ASC) {
+      return SortDirection.ASC;
+    }
+
+    return SortDirection.DESC;
+  };
+
+  const sortBy = (query.sortBy && String(query.sortBy)) || "createdAt";
+  const searchNameTerm = (query.searchNameTerm as string | undefined) || null;
+
+  return { pagination, sortDirection: sortDirection(), sortBy, searchNameTerm };
 };
